ras-detector/requirements.txt
opencv-python
numpy
matplotlib
pyserial
RubikTwoPhase
